feat(paper): add decision and approval service classes
